20240726-1 (401에러 수정 및 검색 기능 추가) commit

+++ hooks/useLogout2.js
@@ -12,6 +12,25 @@ const useLogout = () => {
     setLogoutDialogOpen(false);
   };
 
+  const refreshAccessToken = async (refreshToken) => {
+    try {
+      const response = await axios.post(
+        "http://localhost:8080/admins/refreshToken",
+        {
+          refreshToken,
+        }
+      );
+      if (response.status === 200) {
+        const { accessToken } = response.data;
+        localStorage.setItem("accessToken", accessToken);
+        return accessToken;
+      }
+    } catch (error) {
+      console.error("Failed to refresh access token", error);
+      return null;
+    }
+  };
+
   const handleLogout = async () => {
     const role = localStorage.getItem("role");
     const accessToken = localStorage.getItem("accessToken");
@@ -19,21 +38,53 @@ const useLogout = () => {
 
     if (role && accessToken && refreshToken) {
       try {
-        const response = await axios.delete(
-          `http://localhost:8080/admins/logout`,
-          {
-            headers: {
-              Authorization: `Bearer ${accessToken}`,
-            },
-            data: { refreshToken },
-          }
-        );
+        let token = accessToken;
 
-        if (response.status === 200) {
-          localStorage.removeItem("role");
-          localStorage.removeItem("accessToken");
-          localStorage.removeItem("refreshToken");
-          window.location.href = "/admins/login";
+        try {
+          const response = await axios.delete(
+            `http://localhost:8080/admins/logout`,
+            {
+              headers: {
+                Authorization: `Bearer ${token}`,
+              },
+              data: { refreshToken },
+            }
+          );
+
+          if (response.status === 200) {
+            localStorage.removeItem("role");
+            localStorage.removeItem("accessToken");
+            localStorage.removeItem("refreshToken");
+            window.location.href = "/admins/login";
+          }
+        } catch (error) {
+          if (error.response && error.response.status === 401) {
+            // 액세스 토큰이 만료된 경우
+            token = await refreshAccessToken(refreshToken);
+
+            if (token) {
+              const retryResponse = await axios.delete(
+                `http://localhost:8080/admins/logout`,
+                {
+                  headers: {
+                    Authorization: `Bearer ${token}`,
+                  },
+                  data: { refreshToken },
+                }
+              );
+
+              if (retryResponse.status === 200) {
+                localStorage.removeItem("role");
+                localStorage.removeItem("accessToken");
+                localStorage.removeItem("refreshToken");
+                window.location.href = "/admins/login";
+              }
+            } else {
+              console.error("Failed to refresh access token");
+            }
+          } else {
+            console.error(error);
+          }
         }
       } catch (error) {
         console.error(error);
